Fix build for Windows 10.

+++ src/audiofilters/genericplc.c
@@ -37,7 +37,7 @@ plc_context_t *generic_plc_create_context(int sample_rate) {
 	 *     - the second TRANSITION_DELAY ms are mixed for smooth transition with the begining of arrived frame
 	 */
 	context->continuity_buffer = ms_malloc0(2*sample_rate*sizeof(int16_t)*TRANSITION_DELAY/1000); /* continuity buffer introduce a TRANSITION_DELAY ms delay */
-	context->plc_buffer_len = sample_rate*sizeof(int16_t)*PLC_BUFFER_LEN; /* length in bytes of the plc_buffer */
+	context->plc_buffer_len = (uint16_t)(sample_rate*sizeof(int16_t)*PLC_BUFFER_LEN); /* length in bytes of the plc_buffer */
 	context->plc_buffer = ms_malloc0(context->plc_buffer_len);
 	context->hamming_window = ms_malloc0(sample_rate*PLC_BUFFER_LEN*sizeof(float));
 	context->plc_out_buffer = ms_malloc0(2*context->plc_buffer_len);
